Set buffer cluster level to MONOTONE_CHARACTERS
The default (MONOTONE_GRAPHEMES) will give combining marks the same
cluster values as their bases (and combines also other grapheme
clusters, though not the complete Unicode grapheme cluster spec).

A side effect of the default is that selecting a glyph for a combining
mark will highlight the mark and the base in the Characters List, but
with the new cluster level they will be highlighted separately so one
can tell what the exact character corresponding to the selected glyphs.

There is also the more extreme CHARACTERS level but this one has no
guarantee of monotonicity which FontGoggles currently requires (the
other two level will ensure that clusters always monotonically increase
in the buffer direction even if re-ordering is involved).